+++ lib/src/hooks_provider.dart
@@ -1,7 +1,6 @@
 // coverage:ignore-file
 import 'dart:io';
 
-import 'package:dart_pre_commit/src/tasks/library_imports_task.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:riverpod/riverpod.dart';
 
@@ -10,6 +9,7 @@ import 'task_base.dart';
 import 'tasks/analyze_task.dart';
 import 'tasks/fix_imports_task.dart';
 import 'tasks/format_task.dart';
+import 'tasks/library_imports_task.dart';
 import 'tasks/outdated_task.dart';
 import 'tasks/pull_up_dependencies_task.dart';
 import 'util/file_resolver.dart';