+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1] - 2021-05-07
+### Fixed
+- The `library-imports` can now be properly ignored by adding 
+`// dart_pre_commit:ignore-library-import` before the line where the import 
+happens
+
 ## [2.3.0] - 2021-04-30
 ### Added
 - New Task `library-imports` (#12)